Refactor getCommitsNotSavedInDB method in GetTDDCyclesUseCase
