+++ server/src/modules/TDDCycles/Domain/IDBCommitsRepository.ts
@@ -1,5 +1,6 @@
 import { Pool } from "pg";
 import { CommitDTO } from "./CommitDataObject";
+import { CommitDataObject } from "./commitInterfaces";
 
 export interface IDBCommitsRepository {
   pool: Pool;
@@ -7,4 +8,5 @@ export interface IDBCommitsRepository {
   getCommits(owner: string, repoName: string): Promise<any>;
   commitExists(owner: string, repoName: string, sha: string): Promise<any>;
   repositoryExists(owner: string, repoName: string): Promise<boolean>;
+  getCommitsNotSavedInDB(owner: string, repoName: string, commitsData: CommitDataObject[]): Promise<CommitDataObject[]>;
 }
